+++ src/api/refreshUser.ts
@@ -1,13 +1,19 @@
 import endpoints from '@config/endpoints.js';
-import { useQuery } from 'react-query';
 import apiCall from './apiCall';
+import { useMutation } from 'react-query';
+import { IRefreshUserResponse } from '../types';
 
-const api = async ({ queryKey }: { queryKey: string[] }) => {
-  const [_key, id] = queryKey;
+interface IRefreshUser {
+  id: number;
+}
+
+const api = async (data: IRefreshUser): Promise<IRefreshUserResponse> => {
+  const { id } = data;
+
   const response = await apiCall.get(`${endpoints.REFRESH_SCORE}/${id}`);
 
   return response.data;
 };
 
-const useRefreshUser = ({ id }: { id: string }) => useQuery(['get-refresh-user', id], api);
+const useRefreshUser = () => useMutation(['user-create'], api);
 export default useRefreshUser;