+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
-## 9/10/2024
+## 10/9/2024
+
+- Added analytics tracking code
+- Added Refresh User Score
+- Made Scoreboard into mulitple components for re-usability
+
+## 9/9/2024
 
 - Finished frontend with preact added scoreboard and details view
 - Finished API for create users, get users, get user by id etc
+- Created api endpoints with react query for create, search, get single, refresh and get github user
+- Created search functionality for users
+- Added husky for pre-commit hooks
+- Added eslint for code linting
+- Added prettier for code formatting
+- Added mysql and typeorm for the database
+- Added a rate limiter for the API